refact: page 리펙토링
-BookmarkListItem 컴포넌트로 리펙토링
