+++ src/app/(auth)/(with-header)/bookmarks/[bookmarkGroup]/components/bookmarkDetailPageType.ts
@@ -1,0 +1,12 @@
+export interface IQuestionProps {
+  id: string;
+  createdAt: string;
+  updatedAt: string;
+  title: string;
+  content: string;
+  difficulty: string;
+  likes: number;
+  isSample: boolean;
+  categoryID: string;
+  userID: string;
+}
